Move tests out of the txdbus package
* We dont' want to install tests on the end users' system.
* This excludes tests from the coverage result (which can be deceiving).
* This is kinda the standard practice for python packages.
