+++ setup.py
@@ -20,8 +20,7 @@ setup(
     author_email     = 'tom.cocagne@gmail.com',
     install_requires = ['twisted>=10.1', 'six'],
     provides         = ['txdbus'],
-    packages         = ['txdbus',
-                        'txdbus.test'],
+    packages         = ['txdbus'],
     keywords         = ['dbus', 'twisted'],
     classifiers      = ['Development Status :: 4 - Beta',
                         'Framework :: Twisted',
