+++ tests/client_tests.py
