+++ tox.ini
@@ -6,7 +6,7 @@ skip_missing_interpreters = True
 deps =
   -rrequirements.txt
   coverage
-commands = coverage run --source txdbus -m twisted.trial txdbus/test
+commands = coverage run --source txdbus -m twisted.trial tests
 passenv =
   DBUS_SESSION_BUS_ADDRESS
   USERNAME
